moved helper methods to a module
